[LEAK] minor leak from destructor fixed cause of bst

+++ softheap.cc
@@ -3,45 +3,44 @@
 
 
 template<typename E>
-void SoftHeap<E>::insert(E e)
-{
+void SoftHeap<E>::insert(E e) {
     meld(new SoftHeap<E>(e));
 }
 
 
 template<typename E>
-void swap(E* a, E* b)
-{
+void swap(E *a, E *b) {
     E cpy = *a;
     *a = *b;
-    *b = cpy; 
+    *b = cpy;
 }
 
 
 template<typename E>
-void SoftHeap<E>::meld(SoftHeap* Q)
-{
-    if (this->rank > Q->rank)
-    {
+void SoftHeap<E>::meld(SoftHeap *Q) {
+    if (this->rank > Q->rank) {
         swap(this, Q);
     }
 
-    merge_into(this,Q);
-    repeated_combine(Q,this->rank);
+    merge_into(this, Q);
+    repeated_combine(Q, this->rank);
 
-    auto cpy = this;
-    this = Q;
+    Tree *cpy = this->first;
+    this->epsilon = Q->epsilon;
+    this->rank = Q->rank;
+    this->max_node_rank = Q->max_node_rank;
+
+    this->first = Q->first;
+
     delete cpy;
 }
 
 template<typename E>
-E SoftHeap<E>::pick_elem(Node* x)
-{
-    SoftHeap<E>::ListCell* cpy = x->list;
-    SoftHeap<E>::ListCell* last = x->list;
+E SoftHeap<E>::pick_elem(Node *x) {
+    SoftHeap<E>::ListCell *cpy = x->list;
+    SoftHeap<E>::ListCell *last = x->list;
 
-    while (cpy != nullptr)
-    {
+    while (cpy != nullptr) {
         last = cpy;
         cpy = cpy->next;
     }
@@ -52,23 +51,19 @@ E SoftHeap<E>::pick_elem(Node* x)
 }
 
 template<typename E>
-E SoftHeap<E>::extract_min()
-{
+E SoftHeap<E>::extract_min() {
     if (this->first == nullptr)
         return nullptr;
 
     auto Tree = this->first->sufmin;
-    Node* x = Tree->root;
+    Node *x = Tree->root;
     E e = pick_elem(x);
 
-    if (x->list.size() <= x->size/2)
-    {
-        if (leaf(x) == nullptr)
-        {
+    if (x->list.size() <= x->size / 2) {
+        if (leaf(x) == nullptr) {
             sift(x);
             update_suffix_min(Tree);
-        }
-        else if (x->list == nullptr)
+        } else if (x->list == nullptr)
             remove_tree(this, Tree);
     }
     return e;
@@ -81,7 +76,7 @@ int SoftHeap<E>::ListCell::size() {
     if (list == nullptr)
         return 0;
     int cpt = 1;
-    while(list->next){
+    while (list->next) {
         list = list->next;
         cpt++;
     }
@@ -89,16 +84,16 @@ int SoftHeap<E>::ListCell::size() {
 }
 
 template<typename E>
-void SoftHeap<E>::concatenate(ListCell *l1, ListCell *l2){
+void SoftHeap<E>::concatenate(ListCell *l1, ListCell *l2) {
     if (l1->next == nullptr)
         l1->next = l2;
     else
-        concatenate(l1->next,l2);
+        concatenate(l1->next, l2);
 }
 
 template<typename E>
 void SoftHeap<E>::sift(Node *x) {
-    while(x->list->size() < x->size && !leaf(x)){
+    while (x->list->size() < x->size && !leaf(x)) {
         if (x->left == nullptr || (x->right != nullptr && x->left->ckey > x->right->ckey))
             swap(x->left, x->right);
 
@@ -114,12 +109,12 @@ void SoftHeap<E>::sift(Node *x) {
 }
 
 template<typename E>
-typename SoftHeap<E>::Node* SoftHeap<E>::combine(Node *x, Node *y) {
+typename SoftHeap<E>::Node *SoftHeap<E>::combine(Node *x, Node *y) {
     Node *z = new Node(x, y);
     if (x->rank <= this->max_node_rank)
         z->size = 1;
     else
-        z->size = (3 * x->size + 1)/2;
+        z->size = (3 * x->size + 1) / 2;
     sift(z);
     return z;
 }
@@ -137,8 +132,8 @@ void SoftHeap<E>::merge_into(SoftHeap *q) {
     Tree *t1 = this->first;
     Tree *t2 = q->first;
 
-    while(t1 != nullptr){
-        while(t1->rank > t2->rank)
+    while (t1 != nullptr) {
+        while (t1->rank > t2->rank)
             t2 = t2->next;
         Tree *t1bis = t1->next;
         insert_tree(q, t1, t2);
@@ -157,8 +152,8 @@ void SoftHeap<E>::insert_tree(SoftHeap *q, Tree *t1, Tree *t2) {
 
 template<typename E>
 void SoftHeap<E>::update_suffix_min(Tree *t) {
-    while(t != nullptr) {
-        if(t->root->ckey <= t->next->sufmin->root->ckey)
+    while (t != nullptr) {
+        if (t->root->ckey <= t->next->sufmin->root->ckey)
             t->sufmin = t;
         else
             t->sufmin = t->next->sufmin;
@@ -184,15 +179,13 @@ void SoftHeap<E>::remove_tree(SoftHeap *q, Tree *t) {
 template<typename E>
 void SoftHeap<E>::repeated_combine(SoftHeap *q, int k) {
     Tree *t = q->first;
-    while(t->next != nullptr){
-        if (t->rank == t->next->rank)
-            if (t->next->next == nullptr || t->rank != t->next->next->rank){
-                t->root = combine(t->root, t->next->root);
-                t->rank = t->root->rank;
-                remove_tree(q, t->next);
-            }
-            else if (t->rank > k)
-                break;
+    while (t->next != nullptr) {
+        if (t->next->rank == t->rank && (t->next->next == nullptr || t->rank != t->next->next->rank)) {
+            t->root = combine(t->root, t->next->root);
+            t->rank = t->root->rank;
+            remove_tree(q, t->next);
+        } else if (t->rank > k)
+            break;
         t->next;
     }
 
@@ -202,12 +195,12 @@ void SoftHeap<E>::repeated_combine(SoftHeap *q, int k) {
 }
 
 template<typename E>
-bool SoftHeap<E>::searchAndDestroy(Node *x, E e){
+bool SoftHeap<E>::searchAndDestroy(Node *x, E e) {
     ListCell *l = x->list;
     bool lft = false;
     bool rgt = false;
-    while(l != nullptr){
-        if (l->elem == e){
+    while (l != nullptr) {
+        if (l->elem == e) {
             /////
             return true;
         }
@@ -223,11 +216,17 @@ bool SoftHeap<E>::searchAndDestroy(Node *x, E e){
 template<typename E>
 bool SoftHeap<E>::deleteE(E e) {
     Tree *t = this->first;
-    while(t != nullptr){
+    while (t != nullptr) {
         bool tmp = searchAndDestroy(t->root, e);
         if (tmp)
             return tmp;
         t = t->next;
     }
     return false;
+}
+
+int main() {
+    SoftHeap<int> *s = new SoftHeap<int>(5);
+    std::cout << "CREATE SOFT HEAP WORKS" << std::endl;
+    delete s;
 }
